+++ src/shapes/text.js
@@ -1,5 +1,5 @@
 import _ from "underscore"
-import { Moveable, Styleable, Groupable } from "../mixins"
+import { Moveable, Styleable } from "../mixins"
 import Utils from '../utils'
 
 class Text {
@@ -41,6 +41,6 @@ class Text {
 
 }
 
-_.extend(Text.prototype, Moveable, Styleable, Groupable, { type: "text" });
+_.extend(Text.prototype, Moveable, Styleable, { type: "text" });
 
 export default Text;