+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.9
+
+- `Rune.Polygon` now has a `contains(x,y)` function to know whether a point is inside or outside a polygon
+- Group and shapes now has a `stagepos()` function that returns their absolute position according to the stage
+- Removing `Groupable` mixin in favor of `Moveable`
+
 ## 0.2.8
 
 - Text shapes are not copied with all their variables when calling `copy()`.